Fixed KNearestNeighbor
The function for this was running for way longer than it should have. KNNImputer function is slow across large datasets when filling in missing values. It also only imputed one value across all the missing values, which created a scatterplot line that essentially looked like a straight line. This has now been fixed, the values are generated based on a nearest neighbour of 4, and the time to finish generating and filling in the data is significantly improved, takes seconds as opposed to 5+ minutes.